Use a cross-platform representation for paths
Serializing PathBuf on a machine and deserializing it on a machine witha
different operating system is unsupported because of the way each OS
encodes its paths. With this commit we are introducing the
cross_platform module, which exposes a basic Path and PathComponent
structs that are able to encode OS-specific paths into a single data
type.

This type is then used in fs::Entry and other call sites to speak about
paths in a platform-agnostic way. Theoretically, after this commit, the
only call sites that don't use cross_platform::Path are those ones that
are platform-dependent (e.g., LocalProject::resolve_path) or where
platform doesn't matter (e.g., PathSearchResult::display_path).

Note also that we are only currently dealing with Unix paths, but this
abstraction should allow to handle Windows paths quite easily.